refactorings (created component CustomHubSelection)
